cuboid challenge/requirements.txt
google-generativeai
pandas
matplotlib
plotly
numpy